Move models and crypto to core library
