+++ core/__init__.py
@@ -1,5 +1,6 @@
 """Common library package."""
 
 from .utils import greet
+from .db import db
 
-__all__ = ["greet"]
+__all__ = ["greet", "db"]
