+++ core/db.py
@@ -1,0 +1,7 @@
+from flask_sqlalchemy import SQLAlchemy
+
+# Shared database instance for ORM models
+
+db = SQLAlchemy()
+
+__all__ = ["db"]
