+++ core/models/__init__.py
@@ -1,0 +1,1 @@
+"""ORM models shared across applications."""
