+++ core/models/user.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from flask_login import UserMixin
-from ..extensions import db, login_manager
+from core.db import db
+from webapp.extensions import login_manager
 from werkzeug.security import generate_password_hash, check_password_hash
