+++ webapp/extensions.py
@@ -1,14 +1,14 @@
-from flask_sqlalchemy import SQLAlchemy
+from core.db import db
 from flask_migrate import Migrate
 from flask_login import LoginManager
 from flask_babel import Babel
 from flask_babel import lazy_gettext as _l
 
 
-db = SQLAlchemy()
 migrate = Migrate()
 login_manager = LoginManager()
 login_manager.login_view = "auth.login"  # 未ログイン時のリダイレクト先
 babel = Babel()
 
 login_manager.login_message = None
+
