+++ webapp/photo_view/routes.py
@@ -19,7 +19,7 @@ from flask_babel import gettext as _
 
 from . import bp
 from ..extensions import db
-from ..models.google_account import GoogleAccount
+from core.models.google_account import GoogleAccount
